DEV-11493 update run_reset_database
